Use Mock Decrypt function for WC and Nifty
If wallet is metamask, then provide the eth_decrypt function to the cpinner client. If Nifty or WalletConnect, throw an error that is caught by step2 (previous commit). This only happens if the data is encrypted in the datavault and passed to the wallet to decrypt. This does not happen if the client detects the content is not encrypted.

Edge case: User uses Metamask to insert encrypted data into DV. Then they switch to WalletConnect or Nifty and try to access the DataVault.